Move some files out of the debug folder so they can be reused
